Extract wall preview source helper
